Refactor backupService
