--- sdk/src/main/java/com/microsoft/did/sdk/datasource/file/models/JweProtectedBackup.kt
@@ -1,9 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved
-
-package com.microsoft.did.sdk.datasource.file.models
-
-import com.microsoft.did.sdk.crypto.protocols.jose.jwe.JweToken
-
-abstract class JweProtectedBackup {
-    abstract val jweToken: JweToken
-}
