+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/file/models/ProtectedBackup.kt
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation. All rights reserved
+
+package com.microsoft.did.sdk.datasource.file.models
+
+/**
+ * A ProtectedBackup holds a UnprotectedBackup in some shape or form.
+ * The details are defined by implementations of this class.
+ * e.g. a JWE Token encrypted by a password.
+ */
+abstract class ProtectedBackup
