--- sdk/src/main/java/com/microsoft/did/sdk/datasource/file/models/VcMetadata.kt
@@ -1,11 +1,0 @@
-package com.microsoft.did.sdk.datasource.file.models
-
-import com.microsoft.did.sdk.credential.service.models.contracts.display.DisplayContract
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-@SerialName("BaseVC")
-abstract class VcMetadata {
-    abstract val displayContract: DisplayContract
-}
