--- sdk/src/main/java/com/microsoft/did/sdk/datasource/file/models/WalletMetadata.kt
@@ -1,8 +1,0 @@
-package com.microsoft.did.sdk.datasource.file.models
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-@SerialName("BaseWallet")
-open class WalletMetadata
