+++ sdk/src/main/java/com/microsoft/did/sdk/datasource/file/models/microsoft2020/WalletMetadata.kt
@@ -1,0 +1,8 @@
+package com.microsoft.did.sdk.datasource.file.models.microsoft2020
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+@SerialName("BaseWallet")
+open class WalletMetadata
