+++ sdk/src/test/java/com/microsoft/did/sdk/datasource/file/models/PasswordProtectedJweBackupTest.kt
@@ -3,6 +3,8 @@
 package com.microsoft.did.sdk.datasource.file.models
 
 import com.microsoft.did.sdk.crypto.protocols.jose.jwe.JweToken
+import com.microsoft.did.sdk.datasource.file.models.microsoft2020.Microsoft2020UnprotectedBackupData
+import com.microsoft.did.sdk.datasource.file.models.microsoft2020.WalletMetadata
 import com.microsoft.did.sdk.util.defaultTestSerializer
 import com.nimbusds.jose.EncryptionMethod
 import com.nimbusds.jose.JWEAlgorithm
